tests: gpio-echo: service to support an external latency tester tool
The following service is designed to respond to the latency measure
 requests issued from a remote board.

The code has been validated using rpi3b and a FRDMk64f board running
Zephyr

The Zephyr code is available in this branch (commit for frdmk64f)
https://github.com/ldts/zephyr/commits/evl-latency
https://github.com/ldts/zephyr/blob/evl-latency/samples/drivers/gpio_latency/src/main.c

The DUT running Linux (ie on rpi3b) using GPIOS 23 (to receive) and GPIO24 (to
send) would be running the gpio-echo code  capable of responding to the
requests sent by this latency measurement test code.

gpio-echo is realtime capable (via EVL) or non-realtime capable (standard
linux behviour) - it depends on how you build it.

   $ gpio-echo -n gpiochip0 -o 23 -t 24 -O -T -f

Once that process is started (and the necessary cabling is done) start this
Zephyr program, get the console and follow the instructions.

Connections:
-------------
 Zephyr - FRDMk64F:                Linux - rpi3b
  PIN 20 (PTE-24) ----------------  PIN 16 (GPIO 23)
  PIN 18 (PTE-25) ----------------  PIN 18 (GPIO 24)

Signed-off-by: Jorge Ramirez-Ortiz <jro@xenomai.org>